Updated all module names to PEP 8 standards. Project restructure
