+++ nhanes/variables.py
@@ -1,0 +1,343 @@
+class Tests:
+    seqn = 'SEQN'
+    age = 'RIDAGEYR'
+    crp = 'LBXCRP'
+    hscrp = 'LBXHSCRP'
+    ldh = 'LBXSLDSI'
+    platelets = 'LBXPLTSI'
+    hgb = 'LBXHGB'
+    hct = 'LBXHCT'
+    cd3 = 'LBXCD3'
+    cd4 = 'LBXCD4'
+    cd8 = 'LBXCD8'
+    wbc = 'LBXWBCSI'
+    lymph = 'LBDLYMNO'
+    lymph2 = 'LB2LYMNO'
+    lymphocytes_pct = 'LBXLYPCT'
+    eosinophils = 'LBDEONO'
+    eosinophils2 = 'LB2EONO'
+    eosinophils_pct = 'LBXEOPCT'
+    neutrophils = 'LBDNENO'
+    neutrophils_pct = 'LBXNEPCT'
+    monocytes = 'LBDMONO'
+    monocytes_pct = 'LBXMOPCT'
+    glucose = 'LBXSGL'
+    a1c = 'LBXGH'
+    basophils = 'LBDBANO'
+    basophils_pct = 'LBDBANO'
+    rbc = 'LBXRBCSI'
+    rdw = 'LBXRDW'
+    mpv = 'LBXMPSI'
+    mchc = 'LBXMC'
+    mcv = 'LBXMCVSI'
+    mch = 'LBXMCHSI'
+    alt = 'LBXSATSI'
+    alp = 'LBXSAPSI'
+    ast = 'LBXSASSI'
+    ggt = 'LBXSGTSI'
+    bilirubin = 'LBDSTBSI'
+    cpk = ''  # 'LBXSCK'
+    albumin = 'LBDSALSI'
+    protein = 'LBXSTP'
+    globulin = 'LBDSGBSI'
+    chol = 'LBXSCH'
+    ldl = 'LBDLDL'
+    hdl = 'LBDHDDSI'
+    trigly = 'LBXSTR'
+    creatinin = 'LBDSCRSI'
+    bun = 'LBXSBU'
+    tsh = 'LBXTSH1'
+    ft3 = 'LBXT3F'
+    ft4 = 'LBDT4FSI'
+    tt3 = 'LBXTT3'
+    tt4 = 'LBXTT4'
+    tpoab = 'LBXTPO'
+    tgab = 'LBXATG'
+    tg = 'LBXTGN'
+    ttg = 'LBXTTG'
+    tgema = 'LBXEMA'
+    bicarb = 'LBXSC3SI'
+    calcium = 'LBXSCA'
+    chloride = 'LBXSCLSI'
+    sodium = 'LBXSNASI'
+    potassium = 'LBXSKSI'
+    phosphorus = 'LBXSPH'
+    uric = 'LBDSUASI'
+    b12 = 'LBDB12'
+    vitd = 'LB2VID'
+    folate_rbc = 'LBXRBF'
+    folate = 'LBDFOL'
+    ferritin = 'LBDFER'
+    ferritin_si = 'LBXFERSI'
+    transferin_receptor = 'LBXTFR'
+
+    alcohol_first = 'ALQ010SAS'
+    alcohol_lifetime = 'ALD020'
+    alcohol_past30days = 'ALD030'
+    alcohol_past30days_5 = 'ALD040'
+
+    bpq_ever_told_hypertension = 'BPQ020S'
+    bpq_ever_told_multiple_hypertension = 'BPQ030'
+    bpq_ever_told_hypertension = 'BPQ020SAS'
+    bpq_age_told = 'BPQ035'
+    bpq_taking_prescription = 'BPQ040A'
+    bpq_taking_prescription_current = 'BPQ050A'
+    bpq_ever_told_high_cholesterol = 'BPQ080'
+    bpq_ever_told_cholesterol_reduce_fat = 'BPQ090A'
+    bpq_ever_told_cholesterol_reduce_weight = 'BPQ090B'
+    bpq_ever_told_cholesterol_exercise = 'BPQ090C'
+    bpq_ever_told_take_medicine = 'BPQ090D'
+
+    bowel_leakage_gas = 'BHQ010'
+    bowel_leakage_mucus = 'BHQ020'
+    bowel_leakage_liquid_stools = 'BHQ030'
+    bowel_leakage_solid_stools = 'BHQ040'
+    bowel_frequency = 'BHD050'
+    bowel_common_stool_type = 'BHQ060'
+
+    chest_pain_ever_had = 'CDQ001'
+    chest_pain_walking_uphill = 'CDQ002'
+    chest_pain_walking_ordinary = 'CDQ003'
+    chest_pain_do_you_slowdown = 'CDQ004'
+    chest_pain_standing_relieve = 'CDQ005'
+    chest_pain_relief_time = 'CDQ006'
+    chest_pain_in_right_arm = 'CDQ009A'
+    chest_pain_in_right_chest = 'CDQ009B'
+    chest_pain_in_neck = 'CDQ009C'
+    chest_pain_in_upper_sternum = 'CDQ009D'
+    chest_pain_in_lower_sternum = 'CDQ009E'
+    chest_pain_in_left_chest = 'CDQ009F'
+    chest_pain_in_right_arm = 'CDQ009G'
+    chest_pain_in_epigastric = 'CDQ009H'
+    chest_pain_severe_half_hour = 'CDQ008'
+    chest_pain_in_lower_sternum = 'CDQ009E'
+    chest_breath_shortness_stairs = 'CDQ010'
+
+    health_general = 'HSD010'
+    health_30days = 'HSQ470'
+    health_30days_mental = 'HSQ480'
+    health_30days_inactive_due = 'HSQ490'
+    health_30days_cold = 'HSQ500'
+    health_30days_intestinal = 'HSQ510'
+    health_30days_infection = 'HSQ520'
+    health_12month_blood_donation = 'HSQ571'
+    health_last_blood_donation = 'HSQ580'
+    health_ever_tested_hiv = 'HSQ590'
+
+    skin_ever_told_psoriasis = 'DEQ053'
+    skin_psoriasis_issue = 'DEQ055'
+    skin_psoriasis_little = 'DEQ057'
+
+    diabetes_ever_told = 'DIQ010'
+    diabetes_age_told = 'DIQ040'
+    diabetes_ever_told_prediabetes = 'DIQ160'
+    diabetes_ever_told_risk = 'DIQ170'
+    diabetes_feel_risk = 'DIQ172'
+    diabetes_risk_family_history = 'DIQ175A'
+    diabetes_risk_overweight = 'DIQ175B'
+    diabetes_risk_age = 'DIQ175C'
+    diabetes_risk_poor_diet = 'DIQ175D'
+    diabetes_risk_race = 'DIQ175E'
+    diabetes_risk_baby9lb = 'DIQ175F'
+    diabetes_risk_lack_activity = 'DIQ175G'
+    diabetes_risk_hbp = 'DIQ175H'
+    diabetes_risk_hb_sugar = 'DIQ175I'
+    diabetes_risk_high_cholesterol = 'DIQ175J'
+    diabetes_risk_hypoglycemic = 'DIQ175K'
+    diabetes_risk_extreme_hunger = 'DIQ175L'
+    diabetes_risk_tingling = 'DIQ175M'
+    diabetes_risk_blurred_vision = 'DIQ175N'
+
+    diabetes_risk_increased_fatigue = 'DIQ175O'
+    diabetes_risk_anyone = 'DIQ175P'
+    diabetes_risk_doctor_warning = 'DIQ175Q'
+    diabetes_risk_gestational = 'DIQ175S'
+    diabetes_risk_frequent_urination = 'DIQ175T'
+    diabetes_risk_thirst = 'DIQ175U'
+    diabetes_risk_craving_sugar = 'DIQ175V'
+    diabetes_risk_medication = 'DIQ175W'
+    diabetes_risk_polycystic = 'DIQ175X'
+    diabetes_tested_past_3y = 'DIQ180'
+    diabetes_taking_insulin = 'DIQ050'
+    diabetes_taking_insulin_length = 'DIQ060'
+    diabetes_taking_insulin_unit = 'DIQ060U'
+    diabetes_taking_pills = 'DIQ070'
+    diabetes_last_seen = 'DIQ230'
+    diabetes_frequency_checking = 'DIQ260'
+    diabetes_frequency_unit = 'DIQ260U'
+    diabetes_a1c_past_year = 'DIQ275'
+    diabetes_a1c_last_level = 'DIQ280'
+    diabetes_a1c_should_be = 'DIQ291'
+    diabetes_recent_sbp = 'DIQ300S'
+    diabetes_recent_dbp = 'DIQ300D'
+    diabetes_doctor_sbp = 'DID310S'
+    diabetes_doctor_dbp = 'DID310D'
+    diabetes_recent_ldl = 'DID320'  # error:Q?
+    diabetes_doctor_ldl = 'DID330'
+    diabetes_eyes_affected = 'DIQ080'
+
+    depression_little_interest = 'DPQ010'
+    depression_feeling_down_depressed_hopeless = 'DPQ020'
+    depression_trouble_sleeping = 'DPQ030'
+    depression_feeling_tired = 'DPQ040'
+    depression_poor_too_much_appetite = 'DPQ050'
+    depression_feeling_bad_about_yourself = 'DPQ060'
+    depression_trouble_concentrating = 'DPQ070'
+    depression_being_slow_or_fast = 'DPQ080'
+    depression_thoughts_better_dead = 'DPQ090'
+    depression_how_difficult = 'DPQ100'
+
+    work_type = 'OCD150'
+    work_hours = 'OCQ180'
+    work_full_time = 'OCQ210'
+    work_current_job = 'OCD231'
+    work_job_group = 'OCD241'
+    work_job_situation = 'OCQ260'
+    work_job_months = 'OCD270'
+    work_job_smoking_people = 'OCQ275'
+    work_job_reason_last_week = 'OCQ380'
+    work_longest_job = 'OCD390G'
+    work_longest_job_type = 'OCD391'
+    work_longest_job_group = 'OCD392'
+    work_longest_job_duration = 'OCD395'
+    work_mineral_dust = 'OCQ510'
+    work_mineral_dust_length = 'OCQ520'
+    work_organic_dust = 'OCQ530'
+    work_organic_dust_length = 'OCQ540'
+    work_exhaust_fumes = 'OCQ550'
+    work_exhaust_fumes_length = 'OCQ560'
+    work_fumes_other = 'OCQ570'
+    work_fumes_other_length = 'OCQ580'
+
+    prostate_trouble_starting_urinate = 'KIQ081'
+    prostate_bladder_feel_empty = 'KIQ101'
+    prostate_have_prostate_disease = 'KIQ490'
+    prostate_ever_told_enlarged = 'KIQ121'
+    prostate_benign_enlargement = 'KIQ141'
+    prostate_benign_told_age = 'KIQ161'
+    prostate_enlargement_cancer = 'KIQ182'
+    prostate_rectal_exam = 'KIQ361'
+
+    sleep_hours = "SLD010H"
+    sleep_fall_asleep_time = "SLD020M"
+    sleep_snoring_frequency = "SLQ030"
+    sleep_stop_breathing_frequency = "SLQ040"
+    sleep_ever_told_doctor = "SLQ050"
+    sleep_ever_told_by = 'SLQ060'  # - Ever told by doctor have sleep disorder?
+    sleep_disorder_apnea = 'SLQ070A'  # - Sleep disorder: Sleep Apnea
+    sleep_disorder_insomnia = 'SLQ070B'  # - Sleep disorder: Insomnia
+    sleep_disorder_restless_leg = 'SLQ070C'  # - Sleep disorder: Restless Legs
+    sleep_disorder_other = 'SLQ070D'  # - Sleep disorder: Other
+    sleep_trouble_asleep_frequency = 'SLQ080'  # - How often have trouble falling asleep?
+    sleep_wakeup_frequency = 'SLQ090'  # - How often wake up during night?
+    sleep_wakeup_too_early_frequency = 'SLQ100'  # - How often wake up too early in morning?
+    sleep_feeling_unrest_frequency = 'SLQ110'  # - How often feel unrested during the day?
+    sleep_feel_sleepy_daily_frequency = 'SLQ120'  # - How often feel overly sleepy during day?
+    sleep_not_enough_frequency = 'SLQ130'  # - How often did you not get enough sleep?
+    sleep_pill_frequency = 'SLQ140'  # - How often take pills to help you sleep?
+    sleep_leg_jerks_frequency = 'SLQ150'  # - How often have leg jerks while sleeping?
+    sleep_cramps_frequency = 'SLQ160'  # - How often have legs cramp while sleeping
+    sleep_concentrating_difficulty_tired = 'SLQ170'  # - Difficulty concentrating when tired?
+    sleep_memory_difficulty_tired = 'SLQ180'  # - Difficulty remembering when tired?
+    sleep_eating_difficulty_tired = 'SLQ190'  # - Difficulty eating when tired?
+    sleep_hobby_difficulty_tired = 'SLQ200'  # - Difficulty with a hobby when tired?
+    sleep_things_done_difficulty = 'SLQ210'  # - Difficulty getting things done?
+    sleep_finance_difficulty = 'SLQ220'  # - Difficulty with finance when tired?
+    sleep_work_difficulty = 'SLQ230'  # - Difficulty at work because tired?
+    sleep_phone_difficulty = 'SLQ240'  # - Difficulty on phone when tired?
+
+    smoking_100 = 'SMQ020'  # -Smoked at least 100 cigarettes in life
+    smoking_age_regular = 'SMD030'  # -Age started smoking cigarettes regularly
+    smoking_currently = 'SMQ040'  # -Do you now smoke cigarettes
+    smoking_quit_time = 'SMQ050Q'  # -How long since quit smoking cigarettes
+    smoking_unit = 'SMQ050U'  # -Unit of measure (day/week/month/year)
+    smoking_age_last_regular = 'SMD055'  # -Age last smoked cigarettes regularly
+    smoking_daily_before_quit = 'SMD057'  # -# cigarettes smoked per day when quit
+    smoking_after_waking_time = 'SMQ077'  # -How soon after waking do you smoke
+    smoking_days_in_last_month = 'SMD641'  # -# days smoked cigs during past 30 days
+    smoking_daily_avg = 'SMD650'  # -Avg # cigarettes/day during past 30 days
+    smoking_pack_see = 'SMD093'  # -May I please see the pack of cigarettes
+    smoking_upc = 'SMDUPCA'  # -Cig 12-digit Universal Product Code-UPC
+    smoking_brand = 'SMD100BR'  # -Cigarette Brand/sub-brand
+    smoking_filter = 'SMD100FL'  # -Filter type
+    smoking_menthol = 'SMD100MN'  # -Menthol indicator
+    smoking_length = 'SMD100LN'  # -Length
+    smoking_tar = 'SMD100TR'  # -FTC Tar Content
+    smoking_nicotine = 'SMD100NI'  # -FTC Nicotine Content
+    smoking_co = 'SMD100CO'  # -FTC Carbon Monoxide Content
+    smoking_tried = 'SMQ620'  # -Ever tried cigarette smoking
+    smoking_age_first = 'SMD630'  # -Age first smoked whole cigarette
+    smoking_brands = 'SMQ660'  # -Brands of cigarettes smoked/past mo
+    smoking_malboro = 'SMQ664M'  # -Menthol or non-menthol Marlboro
+    smoking_camels = 'SMQ664C'  # -Menthol or non-menthol Camels
+    smoking_winston = 'SMQ664W'  # -Menthol or non-menthol Winston
+    smoking_bensonhedges = 'SMQ664B'  # -Menthol or non-menthol BensonHedges
+    smoking_other_brands = 'SMQ664O'  # -Menthol or non-menthol other brand
+    smoking_rll_malboro = 'SMQ666M'  # -Regular, light or ultralite Marlboro
+    smoking_rll_camels = 'SMQ666C'  # -Regular, light or ultralite Camels
+    smoking_rll_kools = 'SMQ666K'  # -Regular, light or ultralite Kools
+    smoking_rll_winston = 'SMQ666W'  # -Regular, light or ultralite Winston
+    smoking_rll_bensonhedges = 'SMQ666B'  # -Regular, light or ultralite BensonHedges
+    smoking_rll_salem = 'SMQ666S'  # -Regular, light or ultralite Salem
+    smoking_rll_other = 'SMQ666O'  # -Regular, light or ultralite other brand
+    smoking_tried_quit = 'SMQ670'  # -Tried to quit smoking
+    smoking_qflag = 'SMAQUEX2'  # -Questionnaire Mode Flag
+
+    smoking_household_anyone = 'SMD410'  # - #Does anyone smoking_household = 'SMoke in the home
+    smoking_household_total = 'SMD415'  # - #Total number of smoking_household = 'SMokers in home
+    smoking_household_cigarette = 'SMD415A'  # - #Total # of cigarette smoking_household = 'SMokers in home
+    smoking_household_cigar = 'SMD415B'  # - #Total # of cigar smoking_household = 'SMokers in home
+    smoking_household_pipe = 'SMD415C'  # - #Total # of pipe smoking_household = 'SMokers in home error:typo
+    smoking_household_cigarette_inhome = 'SMD430'  # - #Total # cigarettes smoking_household = 'SMoked in home
+    smoking_household_cigar_inhome = 'SMD440'  # - #Total # cigars smoking_household = 'SMoked in home
+    smoking_household_pipes_inhom = 'SMD450'  # - #Total # pipes smoked in home
+
+    tb_ever_had_skint_test = 'TBQ010'  # - #Ever had TB/tuberculosis skin test
+    tb_ever_had_positive_skin_test = 'TBQ020'  # - #Ever told had positive TB skin test
+    tb_prescribed_preventive_medecine = 'TBQ030'  # - #Prescribed medicine for preventing TB
+    tb_told_active = 'TBQ040'  # - #Ever told you had active TB
+    tb_prescribed_medecine_active = 'TBQ050'  # - #Prescribed medicine for active TB
+    tb_household_active = 'TBQ060'  # - #Lived in household TB sick person
+
+    demo_release_number = 'SDDSRVYR'  # - Data Release Number
+    demo_exam_status = 'RIDSTATR'  # - Interview/Examination Status
+    demo_6_month_period = 'RIDEXMON'  # - Six month time period
+    demo_gender = 'RIAGENDR'  # - Gender
+    demo_age = 'RIDAGEYR'  # - Age at Screening Adjudicated' # - Recode
+    demo_screening_age = 'RIDAGEMN'  # - Age in Months at Screening' # - Recode
+    demo_exam_age = 'RIDAGEEX'  # - Age in Months at Exam' # - Recode
+    demo_ethnicity = 'RIDRETH1'  # - Race/Ethnicity' # - Recode
+    demo_served_forces = 'DMQMILIT'  # - Served in the US Armed Forces
+    demo_birth_country = 'DMDBORN2'  # - Country of Birth' # - Recode
+    demo_citizenship_status = 'DMDCITZN'  # - Citizenship Status
+    demo_time_in_us = 'DMDYRSUS'  # - Length of time in US
+    demo_education_level_child = 'DMDEDUC3'  # - Education Level' # - Children/Youth 6-19
+    demo_education_level_adult = 'DMDEDUC2'  # - Education Level' # - Adults 20+
+    demo_attending_school = 'DMDSCHOL'  # - Now attending school?
+    demo_married = 'DMDMARTL'  # - Marital Status
+    demo_household_number = 'DMDHHSIZ'  # - Total number of people in the Household
+    demo_family_number = 'DMDFMSIZ'  # - Total number of people in the Family
+    demo_household_income = 'INDHHIN2'  # - Annual Household Income
+    demo_family_income = 'INDFMIN2'  # - Annual Family Income
+    demo_poverty_ratio = 'INDFMPIR'  # - Ratio of family income to poverty
+    demo_pregnant = 'RIDEXPRG'  # - Pregnancy Status at Exam' # - Recode
+    demo_r_gender = 'DMDHRGND'  # - HH Ref Person Gender
+    demo_r_age = 'DMDHRAGE'  # - HH Ref Person Age
+    demo_r_birth_country = 'DMDHRBR2'  # - HH Ref Person Country of Birth
+    demo_r_eduction_level = 'DMDHREDU'  # - HH Ref Person Education Level
+    demo_r_marital_status = 'DMDHRMAR'  # - HH Ref Person Marital Status
+    demo_r_spouse_education = 'DMDHSEDU'  # - HH Ref Person's Spouse Education Level
+    demo_language = 'SIALANG'  # - Language of SP Interview
+    demo_proxy = 'SIAPROXY'  # - Proxy used in SP Interview?
+    demo_interpreter = 'SIAINTRP'  # - Interpreter used in SP Interview?
+    demo_family_language = 'FIALANG'  # - Language of Family Interview
+    demo_family_proxy = 'FIAPROXY'  # - Proxy used in Family Interview?
+    demo_family_interpreter = 'FIAINTRP'  # - Interpreter used in Family Interview?
+    demo_mec_language = 'MIALANG'  # - Language of MEC Interview
+    demo_mec_proxy = 'MIAPROXY'  # - Proxy used in MEC Interview?
+    demo_mec_interpreter = 'MIAINTRP'  # - Interpreter used in MEC Interview?
+    demo_acasi_language = 'AIALANG'  # - Language of ACASI Interview
+    demo_weight_interview = 'WTINT2YR'  # - Full Sample 2 Year Interview Weight
+    demo_mec_weight = 'WTMEC2YR'  # - Full Sample 2 Year MEC Exam Weight
+    demo_variance_pseudo_psu = 'SDMVPSU'  # - Masked Variance Pseudo-PSU
+    demo_variance_pseudo_stratum = 'SDMVSTRA'  # - Masked Variance Pseudo-Stratum
